+++ src/felice/consumer.clj
@@ -12,8 +12,8 @@
 
 (defn commit-message-offset [^KafkaConsumer consumer {:keys [partition topic offset]}]
   (let [commit-point (long (inc offset))]
-    (.commit consumer {(TopicPartition. topic partition)
-                       (OffsetAndMetadata. commit-point)})))
+    (.commitSync consumer {(TopicPartition. topic partition)
+                           (OffsetAndMetadata. commit-point)})))
 
 (defn metric->map [^Metric metric]
   (let [metric-name (.metricName metric)]
@@ -133,6 +133,7 @@ to stop the loop.
     (fn []
       (reset! continue? false)
       (deref completion))))
+
 
 
 (defn consumer