+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [0.1.0]
 - basic `producer`
 - basic `consumer` + `process loop`
+- async consumer
 - serialization helpers